Pеализовать двусвязный список для истории просмотров задач

+++ src/tracker/controllers/InMemoryTaskManager.java
@@ -5,6 +5,7 @@ import tracker.model.Subtask;
 import tracker.model.Task;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
 import java.util.HashMap;
@@ -37,6 +38,7 @@ public class InMemoryTaskManager implements TaskManager {
     @Override
     public Task updateTask(Task updatedTask) {
         if (!tasks.containsKey(updatedTask.getId())) {
+            System.out.println("Обновление невозможно. Задача с данным id не существует.");
             return null;
         }
         tasks.put(updatedTask.getId(), updatedTask);
@@ -46,6 +48,7 @@ public class InMemoryTaskManager implements TaskManager {
     @Override
     public Task getTaskById(Integer id) {
         if (!tasks.containsKey(id)) {
+            System.out.println("Задача с данным id не существует.");
             return null;
         }
         historyManager.add(tasks.get(id));
@@ -53,13 +56,38 @@ public class InMemoryTaskManager implements TaskManager {
     }
 
     @Override
-    public void deleteTaskById(Integer id) { tasks.remove(id); }
+    public void deleteTaskById(Integer id) {
+        if (!tasks.containsKey(id)) {
+            System.out.println("Удаление невозможно. Список задач пуст.");
+            return;
+        }
+        historyManager.remove(id);
+        tasks.remove(id);
+    }
 
     @Override
-    public void deleteAllTasks() { tasks.clear(); }
+    public void deleteAllTasks() {
+        if (tasks.isEmpty()) {
+            System.out.println("Удаление невозможно. Список задач пуст.");
+            return;
+        }
+        for (Integer taskId : tasks.keySet()) {
+            historyManager.remove(taskId);
+        }
+        tasks.clear();
+    }
 
     @Override
-    public List<Task> getAllTasks() { return new ArrayList<>(tasks.values()); }
+    public List<Task> getAllTasks() {
+        if (tasks.isEmpty()) {
+            System.out.println("Список задач пуст.");
+            return Collections.emptyList();
+        }
+        for (Task task : tasks.values()) {
+            historyManager.add(task);
+        }
+        return new ArrayList<>(tasks.values());
+    }
 
     @Override
     public int addNewEpic(Epic newEpic) {
@@ -93,31 +121,62 @@ public class InMemoryTaskManager implements TaskManager {
     }
 
     @Override
-    public List<Epic> getAllEpics() { return new ArrayList<>(epicTasks.values()); }
+    public List<Epic> getAllEpics() {
+        if (epicTasks.isEmpty()) {
+            System.out.println("Список эпиков пуст.");
+            return Collections.emptyList();
+        }
+        for (Task epic : epicTasks.values()) {
+            historyManager.add(epic);
+        }
+        return new ArrayList<>(epicTasks.values());
+    }
 
     @Override
     public List<Subtask> getAllEpicSubtasks(int epicId) {
-        Epic epic = epicTasks.get(epicId);
-        return epic.getEpicSubtasks();
+        final Epic epic = epicTasks.get(epicId);
+        final List<Subtask> subs = epic.getEpicSubtasks();
+        if (subs.isEmpty()) {
+            System.out.println("Список подзадач эпика с данным id пуст.");
+            return Collections.emptyList();
+        }
+        for (Task subtask : subs) {
+            historyManager.add(subtask);
+        }
+        return new ArrayList<>(subs);
     }
 
     @Override
     public void deleteEpicById(Integer epicId) {
         if (!epicTasks.containsKey(epicId)) {
+            System.out.println("Удаление невозможно. Список эпиков пуст.");
             return;
         }
         Epic epic = epicTasks.get(epicId);
         List<Subtask> deletedSubs = epic.getEpicSubtasks();
+        epic.deleteAllEpicSubtasks();
         for (Subtask sub : deletedSubs) {
+            historyManager.remove(sub.getId());
             subtasks.remove(sub.getId());
         }
+        historyManager.remove(epicId);
         epicTasks.remove(epicId);
     }
 
     @Override
     public void deleteAllEpics() {
-        epicTasks.clear();
+        if (epicTasks.isEmpty()) {
+            System.out.println("Удаление невозможно. Список эпиков пуст.");
+            return;
+        }
+        for (Integer taskId : subtasks.keySet()) {
+            historyManager.remove(taskId);
+        }
+        for (Integer taskId : epicTasks.keySet()) {
+            historyManager.remove(taskId);
+        }
         subtasks.clear();
+        epicTasks.clear();
     }
 
     @Override
@@ -127,11 +186,10 @@ public class InMemoryTaskManager implements TaskManager {
         }
         int newId = generateNewId();
         newSubtask.setId(newId);
+        newSubtask.setEpicId(epicId);
         subtasks.put(newSubtask.getId(), newSubtask);
         Epic epic = epicTasks.get(epicId);
-        List<Subtask> epicSubtasks = epic.getEpicSubtasks();
-        newSubtask.setEpicId(epicId);
-        epicSubtasks.add(newSubtask);
+        epic.addSubtaskInEpic(newSubtask);
         epic.calculateEpicStatus();
         return newId;
     }
@@ -156,6 +214,7 @@ public class InMemoryTaskManager implements TaskManager {
     @Override
     public Subtask getSubtaskById(Integer subtaskId) {
         if (!subtasks.containsKey(subtaskId)) {
+            System.out.println("Подзадачи с данным id не существует.");
             return null;
         }
         historyManager.add(subtasks.get(subtaskId));
@@ -163,22 +222,40 @@ public class InMemoryTaskManager implements TaskManager {
     }
 
     @Override
-    public List<Subtask> getAllSubtasks() { return new ArrayList<>(subtasks.values()); }
+    public List<Subtask> getAllSubtasks() {
+        if (subtasks.isEmpty()) {
+            System.out.println("Список подзадач пуст.");
+            return Collections.emptyList();
+        }
+        for (Task subtask : subtasks.values()) {
+            historyManager.add(subtask);
+        }
+        return new ArrayList<>(subtasks.values());
+    }
 
     @Override
     public void deleteSubtaskById(Integer id) {
         if (!subtasks.containsKey(id)) {
+            System.out.println("Удаление невозможно. Подзадачи с данным id нет в списке.");
             return;
         }
         Subtask subtask = subtasks.get(id);
         Epic epic = epicTasks.get(subtask.getEpicId());
-        epic.deleteSubtaskInEpic(subtask.getId());
+        epic.deleteSubtaskInEpic(id);
         epic.calculateEpicStatus();
-        subtasks.remove(subtask.getId());
+        historyManager.remove(id);
+        subtasks.remove(id);
     }
 
     @Override
     public void deleteAllSubtasks() {
+        if (subtasks.isEmpty()) {
+            System.out.println("Удаление невозможно. Список подзадач пуст.");
+            return;
+        }
+        for (Integer taskId : subtasks.keySet()) {
+            historyManager.remove(taskId);
+        }
         subtasks.clear();
         for (Epic epic : epicTasks.values()) {
             epic.deleteAllEpicSubtasks();
